requirements.txt
streamlit
streamlit-mic-recorder
librosa
matplotlib
google-generativeai>=0.8.0
numpy
